Modify ghc version CI, add 9.2.1 compatibility
We swapped the ghc version CI to use an ordinary nix-shell, no other
nix integration. This is what we want to ensure that building with
straight up cabal is working (without nix interference). The previous
problem was that doctest would fail without the haskell developPackage
infrastructure. Fortunately, writing out the ghc env vars when building
fixes this problem.