Switch implementation to refined-simple
We want to enforce length > 2 and versions = non-negative, so we
remove instances/functions that could be used to break this (e.g.
IsList). Thus PackageVersion is intentionally more opaque than before.

Furthermore, we switch the implementation from a typical
smart-constructor to refined-simple, as this cuts down on boilerplate
and makes the impl simpler.