Internal rep switched from Int to Word.
We switch the internal representation from Int to Word. This means the
type now preserves both invariants we care about: non-empty and
non-negative.

Because of this, we can replace the unsafe UnsafePackageVersion
constructor with MkPackageVersion (the latter is no longer a pattern
synonym) and promote the unPackageVersion function to a record
selector (to allow OverloadedRecordDot).